+++ src/circularImports.ts
@@ -4,12 +4,18 @@ import { Core, TFileDataType } from './core';
 import { Provider } from './provider';
 import { DEPENDENCY_TYPE, TDependency } from './tdependency';
 import { isCircularImportsEnabled } from './unused-exports/circularImports';
-import { log } from './unused-exports/log';
 import { TNotUsed } from './unused-exports/notUsed';
 
 export class CircularImportsProvider extends Provider {
   constructor(cores: Core[]) {
-    super(cores, areCircularImportsEnabled, TFileDataType.CIRCULAR_IMPORTS, mapFile2Dependency, getNoCircularImports);
+    super(
+      cores,
+      areCircularImportsEnabled,
+      TFileDataType.CIRCULAR_IMPORTS,
+      mapFile2Dependency,
+      getNoCircularImports,
+      false
+    );
   }
 }
 
@@ -43,15 +49,27 @@ function mapFile2Dependency(
   return row;
 }
 
-function circularImportsInFile(pathToPrj: string | undefined, firstPathname: string, node: TDependency): TDependency[] {
+function circularImportsInFile(
+  pathToPrj: string | undefined,
+  firstPathname: string,
+  node: TDependency
+): TDependency[] {
   const mapFn = mapCircularImport2Dependency(pathToPrj, firstPathname, node);
   return node.circularImports?.map(mapFn) ?? [];
 }
 
-function mapCircularImport2Dependency(pathToPrj: string | undefined, firstPathname: string, parent: TDependency) {
+function mapCircularImport2Dependency(
+  pathToPrj: string | undefined,
+  firstPathname: string,
+  parent: TDependency
+) {
   return (circularImport: string, index: number): TDependency => {
     const absFilePath = resolvePath(pathToPrj, circularImport);
-    const nextImport = getNextImport(firstPathname, parent.circularImports, index);
+    const nextImport = getNextImport(
+      firstPathname,
+      parent.circularImports,
+      index
+    );
 
     return new TDependency(
       parent,
@@ -71,13 +89,18 @@ function mapCircularImport2Dependency(pathToPrj: string | undefined, firstPathna
   };
 }
 
-function getNextImport(firstPathname: string, circularImports: string[] | undefined, index: number): string {
+function getNextImport(
+  firstPathname: string,
+  circularImports: string[] | undefined,
+  index: number
+): string {
   if (circularImports === undefined) {
     return '';
   }
 
   index++;
-  const pathname = index < circularImports.length ? circularImports[index] : firstPathname;
+  const pathname =
+    index < circularImports.length ? circularImports[index] : firstPathname;
   return getFileBaseName(pathname);
 }